Latest commit, things work.

+++ lib/providers/product_provider.dart
@@ -15,8 +15,8 @@ class ProductProvider with ChangeNotifier {
   List<ProductModel> _totalProducts = [];
   List<ProductModel> _myProducts = [];
 
-  List<ProductModel> get totalProducts => [..._totalProducts];
-  List<ProductModel> get myProducts => [..._myProducts];
+  List<ProductModel> get totalProducts => _totalProducts;
+  List<ProductModel> get myProducts => _myProducts;
 
   // Get UserProvider (in order to use userId
   void updateUserProvider(UserProvider userProvider) {
@@ -54,67 +54,18 @@ class ProductProvider with ChangeNotifier {
     notifyListeners();
   }
 
-  // Add product to FirebaseFirestore and FirebaseStorage
-  Future<void> editProduct(String title, String description, int price,
-      String category, File? imageFile, File? modelFile,
-      [String? oldImageUrl, String? oldModelUrl]) async {
-    String imageUrl;
-    String modelUrl;
-
-    // if oldImage is not changed then use the oldImageUrl
-    // else delete the old image and upload the new image and save its url
-    if (oldImageUrl != null) {
-      imageUrl = oldImageUrl;
-    } else {
-      await _storageService.deleteFile(oldImageUrl!);
-      imageUrl = await _storageService.uploadImageFile(imageFile!);
-    }
-
-    // if oldModel is not changed then use the oldModel
-    // else delete the old Model and upload the new model and save its url
-    if (oldModelUrl != null) {
-      modelUrl = oldModelUrl;
-    } else {
-      await _storageService.deleteFile(oldModelUrl!);
-      modelUrl = modelFile != null
-          ? await _storageService.uploadModelFile(modelFile)
-          : '';
-    }
-
-    // Create ProductModel object
-    ProductModel product = ProductModel(
-        id: '',
-        title: title,
-        description: description,
-        price: price,
-        category: category,
-        ownerID: _userProvider!.getUserId().toString(),
-        imageUrl: imageUrl,
-        modelUrl: modelUrl);
-
-    String productId = await _fireStoreService.createProduct(product);
-
-    product = product.copyWith(id: productId);
-
-    _totalProducts.add(product);
-    _myProducts.add(product);
-    notifyListeners();
-  }
-
   // Fetch product data from FirebaseFirestore
-  Future<void> fetchProductData(String currentUserID) async {
+  Future<void> fetchAllProducts(String currentUserID) async {
     try {
-      List<ProductModel> fetchedProductData =
-          await _fireStoreService.fetchProductData();
+      List<ProductModel> allProducts =
+          await _fireStoreService.fetchAllProducts();
 
-      _totalProducts = fetchedProductData
-          .where((element) =>
-              element.ownerID != currentUserID)
+      _totalProducts = allProducts
+          .where((product) => product.ownerID != currentUserID)
           .toList();
-           
-      _myProducts = fetchedProductData
-          .where((element) =>
-              element.ownerID == currentUserID)
+
+      _myProducts = allProducts
+          .where((product) => product.ownerID == currentUserID)
           .toList();
 
       notifyListeners();
@@ -126,16 +77,26 @@ class ProductProvider with ChangeNotifier {
 
   // Delete product from FirebaseFirestore and FirebaseStorage
   Future<void> deleteProduct(String productId) async {
+    var imageUrl =
+        _myProducts.firstWhere((element) => element.id == productId).imageUrl;
+    var modelUrl =
+        _myProducts.firstWhere((element) => element.id == productId).modelUrl;
     try {
-      // Delete product image from FirebaseStorage
-      await _storageService.deleteFile(_myProducts
-          .firstWhere((element) => element.id == productId)
-          .imageUrl);
+      // Check if the image in product are needed by order documents
+      // If not, delete them from firebase Storage.
+      if (await _fireStoreService.isImageNeededIn(imageUrl, 'orders')) {
+        await _storageService.deleteFile(_myProducts
+            .firstWhere((element) => element.id == productId)
+            .imageUrl);
+      }
 
-      // Delete product 3Dmodel from FirebaseStorage
-      await _storageService.deleteFile(_myProducts
-          .firstWhere((element) => element.id == productId)
-          .modelUrl);
+      // Check if the model in product are needed by order documents
+      // If not, delete them from firebase Storage.
+      if (await _fireStoreService.isModelNeededIn(modelUrl, 'orders') && modelUrl != '') {
+        await _storageService.deleteFile(_myProducts
+            .firstWhere((element) => element.id == productId)
+            .modelUrl);
+      }
 
       // Delete product from FirebaseFirestore
       await _fireStoreService.deleteProduct(productId);
